[move-mutator] mutant verification
This commit introduces mutant verification.

Verification is done by copying the whole package to the temp directory
(which should be cheap as temp directory should reside in some tmpfs so in
RAM generally) and then, original file is replaced by the mutated source.

Compiler is run over the whole package and result is checked. If compilation
fails mutant is not created in the output directory and is not added to
the report.