[move-mutator] use package path correctly and create proper Options
We need to implement defaults manually for the cli::Options just because we need to specify the default value for out_mutant_dir.
Otherwise, out_mutant_dir would be empty. This is special case, when user won't specify any Options (so the default value would be used),
but define package_path (which is passed using other mechanism).

Package path is always set by the mutator caller. It's set to either -p option (by the user) or to ".". We prefer to use
move_sources (-m option) as user may want to pass single Move files, which may be placed somewhere outside any valid package. And
mutator tool should handle them smoothly (those files just can't be verified as we can't compile something that's not a package).