[move-mutator] switch to the AST that comes from `typing` module
This commit switches mutator to the AST that comes from the `typing` compiler module.

Rationale:
`typing` tree is still close enough to the source code, which makes it easy for the mutation
operators to manipulate the source code directly. However, `typing` module gives information
about the expression and literal types.

It occured to be useful to have all expressions parsed and typed, instead of typing particular
parts by the mutation operators themselves.