[move-mutator] prepare deps and resolve names during AST generation
When AST is generated from the `typing` module there is a need to resolve all the deps and names for the
compiler to operate properly. `typing` is the fourth stage so it means that source code must pass through
parser, expansion and naming modules before.

This commit introduces proper compiler preparation for packages and plain Move files.